# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Стильный призыв API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List admins/team (public)
      responses:
        "200":
          description: Admins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAdminsPayload"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Promote an existing Clerk user to admin (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAdminPayload"
  /admins/{id}/role:
    patch:
      operationId: updateAdminRole
      tags: [admins]
      summary: Change an admin's role, optionally transferring ownership (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
  /admins/{id}:
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Remove an admin (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /admins/clerk-users/{id}/ban:
    post:
      operationId: banClerkUser
      tags: [admins]
      summary: Ban a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessPayload"
  /admins/clerk-users/{id}/unban:
    post:
      operationId: unbanClerkUser
      tags: [admins]
      summary: Unban a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessPayload"
  /admins/clerk-users/{id}:
    delete:
      operationId: deleteClerkUser
      tags: [admins]
      summary: Delete a user account (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessPayload"

  # ── Messages (contact inbox) ───────────────────────────────────────────
...
  /chat/users:
    get:
      operationId: listChatUsers
      tags: [chat]
      summary: Search app users to start a chat with
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListChatUsersPayload"
...
  /chat/conversations/support:
    post:
      operationId: startSupportConversation
      tags: [chat]
      summary: Start or fetch the current user's support conversation
      responses:
        "200":
          description: Existing conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartSupportConversationPayload"
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartSupportConversationPayload"
  /chat/conversations/direct:
    post:
      operationId: startDirectConversation
      tags: [chat]
      summary: Start or fetch a direct conversation with another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartDirectConversationInput"
      responses:
        "200":
          description: Existing conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartDirectConversationPayload"
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartDirectConversationPayload"
  /chat/conversations/{id}/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListChatMessagesPayload"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendChatMessagePayload"
  /chat/conversations/{id}/read:
    patch:
      operationId: markConversationRead
      tags: [chat]
      summary: Mark a conversation as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Marked read

  # ── Storage ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessPayload:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # ── Admin / Me ─────────────────────────────────────────────────────
    Admin:
      type: object
      required: [id, clerkUserId, email, name, role, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        email: { type: string }
        name: { type: string }
        role:
          type: string
          enum: [owner, editor]
        createdAt: { type: string }
...
    ListAdminsPayload:
      type: array
      items:
        $ref: "#/components/schemas/Admin"
    AdminInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
        role:
          type: string
          enum: [owner, editor]
    CreateAdminPayload:
      $ref: "#/components/schemas/Admin"
    AdminRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, editor]
        transferOwnership:
          type: boolean
...
    ChatUser:
      type: object
      required: [id, clerkUserId, name, email, updatedAt, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        name: { type: string }
        email: { type: string }
        avatarUrl: { type: ["string", "null"] }
        updatedAt: { type: string }
        createdAt: { type: string }
    ListChatUsersPayload:
      type: array
      items:
        $ref: "#/components/schemas/ChatUser"
...
    StartSupportConversationPayload:
      type: object
      required: [id]
      properties:
        id: { type: integer }
    StartDirectConversationInput:
      type: object
      required: [targetClerkId]
      properties:
        targetClerkId: { type: string }
    StartDirectConversationPayload:
      type: object
      required: [id]
      properties:
        id: { type: integer }
    ChatMessage:
      type: object
      required: [id, conversationId, senderClerkId, senderName, senderIsAdmin, createdAt]
      properties:
        id: { type: integer }
        conversationId: { type: integer }
        senderClerkId: { type: string }
        senderName: { type: string }
        senderAvatarUrl: { type: ["string", "null"] }
        senderIsAdmin:
          type: string
          enum: ["true", "false"]
        content: { type: ["string", "null"] }
        attachmentUrl: { type: ["string", "null"] }
        attachmentType:
          type: ["string", "null"]
          enum: [image, video, file, null]
        attachmentName: { type: ["string", "null"] }
        attachmentMimeType: { type: ["string", "null"] }
        attachmentSize: { type: ["integer", "null"] }
        createdAt: { type: string }
    ListChatMessagesPayload:
      type: array
      items:
        $ref: "#/components/schemas/ChatMessage"
    ChatMessageInput:
      type: object
      properties:
        content: { type: string }
        attachmentUrl: { type: string }
        attachmentType:
          type: string
          enum: [image, video, file]
        attachmentName: { type: string }
        attachmentMimeType: { type: string }
        attachmentSize: { type: integer }
    SendChatMessagePayload:
      $ref: "#/components/schemas/ChatMessage"

    # ── Storage ────────────────────────────────────────────────────────
...