# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Стильный призыв API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRole } from './adminRole';

export interface Admin {
  id: number;
  clerkUserId: string;
  email: string;
  name: string;
  role: AdminRole;
  createdAt: string;
}