# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Стильный призыв API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageAttachmentType } from './chatMessageAttachmentType';
import type { ChatMessageSenderIsAdmin } from './chatMessageSenderIsAdmin';

export interface ChatMessage {
  id: number;
  conversationId: number;
  senderClerkId: string;
  senderName: string;
  /** @nullable */
  senderAvatarUrl?: string | null;
  senderIsAdmin: ChatMessageSenderIsAdmin;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  attachmentUrl?: string | null;
  /** @nullable */
  attachmentType?: ChatMessageAttachmentType;
  /** @nullable */
  attachmentName?: string | null;
  /** @nullable */
  attachmentMimeType?: string | null;
  /** @nullable */
  attachmentSize?: number | null;
  createdAt: string;
}